lambda/requirements.txt
pyjwt
python-dotenv
requests